Cleanup event handlers on destroy
Depending on how the Vue-Redactor component is rendered, it will cause a memory leak. This is because the event handlers for Redactor are not removed when the Vue component is destroyed. Vue is completely unaware of any event handlers that are built from Redactor, so we need to tell Redactor to tear down it's event handlers when removing the Vue component.